+++ src/client/proxy_metric_client.go
@@ -9,8 +9,8 @@ import (
 	"net/http"
 	"strings"
 
-	"github.com/simelo/rextporter/src/common"
 	"github.com/simelo/rextporter/src/config"
+	"github.com/simelo/rextporter/src/util"
 )
 
 // ProxyMetricClient implements the getRemoteInfo method from `client.Client` interface by using some `.toml` config parameters
@@ -22,9 +22,9 @@ type ProxyMetricClient struct {
 
 // NewProxyMetricClient will put all the required info to be able to do http requests to get the remote data.
 func NewProxyMetricClient(service config.Service) (client *ProxyMetricClient, err error) {
-	const generalScopeErr = "error creating a proxy client to get the metrics from remote endpoint"
+	const generalScopeErr = "error creating a forward_metrics client to get the metrics from remote endpoint"
 	if strings.Compare(service.Mode, config.ServiceTypeProxy) != 0 {
-		return client, errors.New("can not create a proxy metric client from a service of type " + service.Mode)
+		return client, errors.New("can not create a forward_metrics metric client from a service of type " + service.Mode)
 	}
 	client = new(ProxyMetricClient)
 	client.BaseClient.service = service
@@ -32,7 +32,7 @@ func NewProxyMetricClient(service config.Service) (client *ProxyMetricClient, er
 	client.BaseClient.req, err = http.NewRequest("GET", service.URIToGetExposedMetric(), nil)
 	if err != nil {
 		errCause := fmt.Sprintln("can not create the request: ", err.Error())
-		return nil, common.ErrorFromThisScope(errCause, generalScopeErr)
+		return nil, util.ErrorFromThisScope(errCause, generalScopeErr)
 	}
 	return client, nil
 }
@@ -43,20 +43,27 @@ func (client *ProxyMetricClient) getRemoteInfo() (data []byte, err error) {
 	var resp *http.Response
 	if resp, err = httpClient.Do(client.req); err != nil {
 		errCause := fmt.Sprintln("can not do the request: ", err.Error())
-		return nil, common.ErrorFromThisScope(errCause, generalScopeErr)
+		return nil, util.ErrorFromThisScope(errCause, generalScopeErr)
 	}
 	if resp.StatusCode != http.StatusOK {
 		errCause := fmt.Sprintf("no success response, status %s", resp.Status)
-		return nil, common.ErrorFromThisScope(errCause, generalScopeErr)
+		return nil, util.ErrorFromThisScope(errCause, generalScopeErr)
 	}
 	defer resp.Body.Close()
 	var reader io.ReadCloser
+	var isGzipContent = false
+	defer func() {
+		if isGzipContent {
+			reader.Close()
+		}
+	}()
 	switch resp.Header.Get("Content-Encoding") {
 	case "gzip":
+		isGzipContent = true
 		reader, err = gzip.NewReader(resp.Body)
 		if err != nil {
 			errCause := fmt.Sprintln("can not create gzip reader.", err.Error())
-			return nil, common.ErrorFromThisScope(errCause, generalScopeErr)
+			return nil, util.ErrorFromThisScope(errCause, generalScopeErr)
 		}
 		defer reader.Close()
 	default:
@@ -65,7 +72,7 @@ func (client *ProxyMetricClient) getRemoteInfo() (data []byte, err error) {
 	// FIXME(denisacostaq@gmail.com): write an integration test for plain text and compressed content
 	if data, err = ioutil.ReadAll(reader); err != nil {
 		errCause := fmt.Sprintln("can not read the body: ", err.Error())
-		return nil, common.ErrorFromThisScope(errCause, generalScopeErr)
+		return nil, util.ErrorFromThisScope(errCause, generalScopeErr)
 	}
 	return data, nil
 }
@@ -74,7 +81,7 @@ func (client *ProxyMetricClient) getRemoteInfo() (data []byte, err error) {
 func (client *ProxyMetricClient) GetExposedMetrics() (data []byte, err error) {
 	const generalScopeErr = "error getting metrics data"
 	if data, err = client.getRemoteInfo(); err != nil {
-		return data, common.ErrorFromThisScope(err.Error(), generalScopeErr)
+		return data, util.ErrorFromThisScope(err.Error(), generalScopeErr)
 	}
 	return data, nil
 }